HBX-1676: Move class 'org.hibernate.tool.hbmlint.detector.ShadowedIdentifierDetector' to 'org.hibernate.tool.internal.export.lint.ShadowedIdentifierDetector'
