+++ orm/src/main/java/org/hibernate/tool/internal/export/lint/ShadowedIdentifierDetector.java
@@ -1,10 +1,7 @@
-package org.hibernate.tool.hbmlint.detector;
+package org.hibernate.tool.internal.export.lint;
 
 import org.hibernate.mapping.PersistentClass;
 import org.hibernate.mapping.Property;
-import org.hibernate.tool.internal.export.lint.EntityModelDetector;
-import org.hibernate.tool.internal.export.lint.Issue;
-import org.hibernate.tool.internal.export.lint.IssueCollector;
 
 public class ShadowedIdentifierDetector extends EntityModelDetector {
